dependencies: Fix structure of Jagex's jar files to conform to standards
`JarInputStream` makes assumptions about the structure of jar files
which Jagex's jars do not conform to. We have to perform some work to
restructure those files to the standard structure.

Ideally this is something we will need to perform quite rarely since
this project will primarly be used by myself with custom
browsercontrol, loader, and client jars.

Signed-off-by: Open592 Developer <uss@open592.com>