dependencies: Fix bug in progress reporting
When calculating the progress we weren't breaking out of the buffer read loop
when we encountered the end of the buffer. Since read() returns -1 when it
encounters the end of the buffer we were removing a byte from
`totalDownloadedBytes` and thus reporting 99% progress even after fully
downloading all remote dependencies.

Signed-off-by: Open592 Developer <uss@open592.com>